test2

+++ Program.cs
@@ -2,7 +2,7 @@
 #region Học phép toán cơ bản
 // Học + - * / %
 // Khai báo biến
-int add = 5 + 10;
+double add = 5 + 10;
 Console.WriteLine($"Kết quả cộng 2 số 5 và 10 là: {add}");
 int minus = 12345 -1234;
 Console.WriteLine($"Kết quả của phép trừ 12345 và 1234 là: {minus}");
